add CSS base + /home display

+++ src/Controller/ProfileArticlesController.php
@@ -9,6 +9,7 @@ use App\Service\FileUploader;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Routing\Annotation\Route;
 
 #[Route('/profile/articles')]
@@ -50,39 +51,51 @@ class ProfileArticlesController extends AbstractController
     #[Route('/{id}', name: 'app_profile_articles_show', methods: ['GET'])]
     public function show(Articles $article): Response
     {
-        return $this->render('profile_articles/show.html.twig', ['article' => $article,]);
+        if ($article->getUser() !== $this->getUser()) {
+            throw $this->createAccessDeniedException('Vous n\'avez pas accès à cet article.');
+        } else {
+            return $this->render('profile_articles/show.html.twig', [
+                'article' => $article,
+            ]);
+        }
     }
 
     #[Route('/{id}/edit', name: 'app_profile_articles_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, FileUploader $fileUploader, Articles $article, ArticlesRepository $articlesRepository): Response
     {
-        $form = $this->createForm(ArticlesType::class, $article);
-        $form->handleRequest($request);
+        if ($article->getUser() !== $this->getUser()) {
+            throw $this->createAccessDeniedException('Vous n\'avez pas accès à cet article.');
+        } else {
+            $form = $this->createForm(ArticlesType::class, $article);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
+            if ($form->isSubmitted() && $form->isValid()) {
 
-            $imageFile = $form->get('image')->getData();
-            if ($imageFile) {
-                $imageFileName = $fileUploader->upload($imageFile);
-                $article->setImageFile($imageFileName);
+                $imageFile = $form->get('image')->getData();
+                if ($imageFile) {
+                    $imageFileName = $fileUploader->upload($imageFile);
+                    $article->setImageFile($imageFileName);
+                }
+
+                $articlesRepository->save($article, true);
+
+                return $this->redirectToRoute('app_profile_articles_index', [], Response::HTTP_SEE_OTHER);
             }
 
-            $articlesRepository->save($article, true);
-
-            return $this->redirectToRoute('app_profile_articles_index', [], Response::HTTP_SEE_OTHER);
+            return $this->renderForm('profile_articles/edit.html.twig', [
+                'article' => $article,
+                'form' => $form,
+            ]);
         }
-
-        return $this->renderForm('profile_articles/edit.html.twig', [
-            'article' => $article,
-            'form' => $form,
-        ]);
     }
 
     #[Route('/{id}', name: 'app_profile_articles_delete', methods: ['POST'])]
     public function delete(Request $request, Articles $article, ArticlesRepository $articlesRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$article->getId(), $request->request->get('_token'))) {
-            $articlesRepository->remove($article, true);
+        if ($article->getUser() === $this->getUser()) {
+            if ($this->isCsrfTokenValid('delete'.$article->getId(), $request->request->get('_token'))) {
+                $articlesRepository->remove($article, true);
+            }
         }
 
         return $this->redirectToRoute('app_profile_articles_index', [], Response::HTTP_SEE_OTHER);
